requirements.txt
annotated-types==0.6.0
anyio==3.7.1
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
boto3==1.24.24
botocore==1.27.24
certifi==2023.11.17
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
commonmark==0.9.1
customtkinter==5.1.3
darkdetect==0.8.0
decorator==4.4.2
dill==0.3.5.1
distlib==0.3.6
dlib==19.24.2
dnspython==2.4.2
EbookLib==0.18
email-validator==2.1.0.post1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
exrex==0.11.0
fastapi==0.104.0
filelock==3.12.0
flatbuffers==23.5.26
fsspec==2023.12.0
future==0.18.2
greenlet==1.1.2
gTTS==2.2.4
h11==0.14.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
humanfriendly==10.0
idna==3.3
imageio==2.31.5
imageio-ffmpeg==0.4.7
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
keyboard==0.13.5
langdetect==1.0.9
lazy_loader==0.3
llvmlite==0.41.0
loguru==0.6.0
lxml==4.9.1
markdown2==2.4.12
MarkupSafe==2.1.3
mido==1.2.10
moviepy==1.0.3
mpmath==1.3.0
multiprocess==0.70.13
networkx==3.1
nltk==3.7
numba==0.58.0
numpy==1.23.1
oauthlib==3.2.0
onnxruntime==1.16.0
opencv-python-headless==4.8.1.78
openpyxl==3.0.10
orjson==3.9.9
outcome==1.2.0
packaging==23.2
pandas==1.4.4
pathos==0.2.9
Pillow==9.1.1
platformdirs==3.5.0
playwright==1.23.0
pooch==1.7.0
pox==0.3.1
ppft==1.7.6.5
praw==7.6.0
prawcore==2.3.0
proglog==0.1.10
protobuf==4.24.3
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
pyee==8.1.0
PyExecJS==1.5.1
pygame==2.1.2
Pygments==2.12.0
PyMatting==1.1.8
pyobjc==8.5
pyobjc-core==8.5
pyobjc-framework-Accessibility==8.5
pyobjc-framework-Accounts==8.5
pyobjc-framework-AddressBook==8.5
pyobjc-framework-AdServices==8.5
pyobjc-framework-AdSupport==8.5
pyobjc-framework-AppleScriptKit==8.5
pyobjc-framework-AppleScriptObjC==8.5
pyobjc-framework-ApplicationServices==8.5
pyobjc-framework-AppTrackingTransparency==8.5
pyobjc-framework-AudioVideoBridging==8.5
pyobjc-framework-AuthenticationServices==8.5
pyobjc-framework-AutomaticAssessmentConfiguration==8.5
pyobjc-framework-Automator==8.5
pyobjc-framework-AVFoundation==8.5
pyobjc-framework-AVKit==8.5
pyobjc-framework-BusinessChat==8.5
pyobjc-framework-CalendarStore==8.5
pyobjc-framework-CallKit==8.5
pyobjc-framework-CFNetwork==8.5
pyobjc-framework-ClassKit==8.5
pyobjc-framework-CloudKit==8.5
pyobjc-framework-Cocoa==8.5
pyobjc-framework-Collaboration==8.5
pyobjc-framework-ColorSync==8.5
pyobjc-framework-Contacts==8.5
pyobjc-framework-ContactsUI==8.5
pyobjc-framework-CoreAudio==8.5
pyobjc-framework-CoreAudioKit==8.5
pyobjc-framework-CoreBluetooth==8.5
pyobjc-framework-CoreData==8.5
pyobjc-framework-CoreHaptics==8.5
pyobjc-framework-CoreLocation==8.5
pyobjc-framework-CoreMedia==8.5
pyobjc-framework-CoreMediaIO==8.5
pyobjc-framework-CoreMIDI==8.5
pyobjc-framework-CoreML==8.5
pyobjc-framework-CoreMotion==8.5
pyobjc-framework-CoreServices==8.5
pyobjc-framework-CoreSpotlight==8.5
pyobjc-framework-CoreText==8.5
pyobjc-framework-CoreWLAN==8.5
pyobjc-framework-CryptoTokenKit==8.5
pyobjc-framework-DataDetection==8.5
pyobjc-framework-DeviceCheck==8.5
pyobjc-framework-DictionaryServices==8.5
pyobjc-framework-DiscRecording==8.5
pyobjc-framework-DiscRecordingUI==8.5
pyobjc-framework-DiskArbitration==8.5
pyobjc-framework-DVDPlayback==8.5
pyobjc-framework-EventKit==8.5
pyobjc-framework-ExceptionHandling==8.5
pyobjc-framework-ExecutionPolicy==8.5
pyobjc-framework-ExternalAccessory==8.5
pyobjc-framework-FileProvider==8.5
pyobjc-framework-FileProviderUI==8.5
pyobjc-framework-FinderSync==8.5
pyobjc-framework-FSEvents==8.5
pyobjc-framework-GameCenter==8.5
pyobjc-framework-GameController==8.5
pyobjc-framework-GameKit==8.5
pyobjc-framework-GameplayKit==8.5
pyobjc-framework-ImageCaptureCore==8.5
pyobjc-framework-IMServicePlugIn==8.5
pyobjc-framework-InputMethodKit==8.5
pyobjc-framework-InstallerPlugins==8.5
pyobjc-framework-InstantMessage==8.5
pyobjc-framework-Intents==8.5
pyobjc-framework-IntentsUI==8.5
pyobjc-framework-IOSurface==8.5
pyobjc-framework-iTunesLibrary==8.5
pyobjc-framework-KernelManagement==8.5
pyobjc-framework-LatentSemanticMapping==8.5
pyobjc-framework-LaunchServices==8.5
pyobjc-framework-libdispatch==8.5
pyobjc-framework-LinkPresentation==8.5
pyobjc-framework-LocalAuthentication==8.5
pyobjc-framework-LocalAuthenticationEmbeddedUI==8.5
pyobjc-framework-MailKit==8.5
pyobjc-framework-MapKit==8.5
pyobjc-framework-MediaAccessibility==8.5
pyobjc-framework-MediaLibrary==8.5
pyobjc-framework-MediaPlayer==8.5
pyobjc-framework-MediaToolbox==8.5
pyobjc-framework-Metal==8.5
pyobjc-framework-MetalKit==8.5
pyobjc-framework-MetalPerformanceShaders==8.5
pyobjc-framework-MetalPerformanceShadersGraph==8.5
pyobjc-framework-MetricKit==8.5
pyobjc-framework-MLCompute==8.5
pyobjc-framework-ModelIO==8.5
pyobjc-framework-MultipeerConnectivity==8.5
pyobjc-framework-NaturalLanguage==8.5
pyobjc-framework-NetFS==8.5
pyobjc-framework-Network==8.5
pyobjc-framework-NetworkExtension==8.5
pyobjc-framework-NotificationCenter==8.5
pyobjc-framework-OpenDirectory==8.5
pyobjc-framework-OSAKit==8.5
pyobjc-framework-OSLog==8.5
pyobjc-framework-PassKit==8.5
pyobjc-framework-PencilKit==8.5
pyobjc-framework-Photos==8.5
pyobjc-framework-PhotosUI==8.5
pyobjc-framework-PreferencePanes==8.5
pyobjc-framework-PushKit==8.5
pyobjc-framework-Quartz==8.5
pyobjc-framework-QuickLookThumbnailing==8.5
pyobjc-framework-ReplayKit==8.5
pyobjc-framework-SafariServices==8.5
pyobjc-framework-SceneKit==8.5
pyobjc-framework-ScreenCaptureKit==8.5
pyobjc-framework-ScreenSaver==8.5
pyobjc-framework-ScreenTime==8.5
pyobjc-framework-ScriptingBridge==8.5
pyobjc-framework-SearchKit==8.5
pyobjc-framework-Security==8.5
pyobjc-framework-SecurityFoundation==8.5
pyobjc-framework-SecurityInterface==8.5
pyobjc-framework-ServiceManagement==8.5
pyobjc-framework-ShazamKit==8.5
pyobjc-framework-Social==8.5
pyobjc-framework-SoundAnalysis==8.5
pyobjc-framework-Speech==8.5
pyobjc-framework-SpriteKit==8.5
pyobjc-framework-StoreKit==8.5
pyobjc-framework-SyncServices==8.5
pyobjc-framework-SystemConfiguration==8.5
pyobjc-framework-SystemExtensions==8.5
pyobjc-framework-UniformTypeIdentifiers==8.5
pyobjc-framework-UserNotifications==8.5
pyobjc-framework-UserNotificationsUI==8.5
pyobjc-framework-VideoSubscriberAccount==8.5
pyobjc-framework-VideoToolbox==8.5
pyobjc-framework-Virtualization==8.5
pyobjc-framework-Vision==8.5
pyobjc-framework-WebKit==8.5
pyperclip==1.8.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
python-rtmidi==1.4.9
python-twitter==3.5
pyttsx3==2.90
pytube==12.1.0
pytz==2022.2.1
PyWavelets==1.4.1
PyYAML==6.0.1
regex==2022.8.17
rembg==2.0.50
requests==2.28.1
requests-futures==1.0.1
requests-oauthlib==1.3.1
rich==12.5.1
s3transfer==0.6.0
scikit-image==0.21.0
scipy==1.11.3
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
starlette==0.27.0
stem==1.8.2
sympy==1.12
textblob==0.17.1
tifffile==2023.9.26
toml==0.10.2
torch==2.2.0.dev20231203
torchaudio==2.2.0.dev20231203
torchvision==0.17.0.dev20231203
torrequest==0.1.0
tqdm==4.64.0
translators==5.3.1
trio==0.22.0
trio-websocket==0.10.2
tweepy==4.10.1
typing_extensions==4.8.0
ujson==5.8.0
undetected-playwright==0.2.0
update-checker==0.18.0
urllib3==1.26.11
uvicorn==0.23.2
uvloop==0.19.0
virtualenv==20.23.0
watchfiles==0.21.0
websocket-client==1.3.3
websockets==10.1
wsproto==1.2.0
